+++ CHANGELOG.md
@@ -18,6 +18,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - `:archive` now works on servers using a different delimiter
 
+### Changed
+
+- Composing an email is now aborted if the text editor exits with an error
+  (e.g. with `vim`, abort an email with `:cq`).
+
 ## [0.15.2](https://git.sr.ht/~rjarry/aerc/refs/0.15.2) - 2023-05-11
 
 ### Fixed